Split data.userconf from configuration.lua (jd.conf).
data.userconf can change and thus must be saved, so that a copy in e.g.
%APPDATA% (on Windows) overrides the settings in the game's file. But before
this split, settings like the number of layers of the DrawService were also
saved which would make it impossible to change them in later updates. Now a
later revision of the game can happily change this settings without having them
overridden by user configuration.

Currently the persistence module is used to  save the copy, C++-Code was not
changed. This has the drawback of duplicating the file/module name, but maybe
here KISS is more important than the small violation of DRY (modules which use
data.userconf must know the name anyway to 'require' them, so having to specify
'userconf' at save time seems not to be a real problem.