Put the code under BSD 2-Clause, artwork under CC BY-SA 3.0.
Add missing information about the license of Ubuntu-R.ttf.
Add full texts of all licenses in the licenses subdirectory.
Move the contests of game/res/snd/sources.txt to LICENSE.txt and translate it
from German to English.
Make a copy in RTF for better readability and iclude it and the licenses
subdirectory in both the ZIP and installer distribution.
Make the installer display the license overview at install time.
Add a short copyright/license notice to the top of all code files.